Create Exercise-7-GuessingGame
Write a program named GuessingGame that generates a random number
between 1 and 10. 
(In other words, in the example above, min is 1 and max is 10)
Ask a user to guess the random number, then display the random number and a
message indicating whether the user’s guess was too high, too low, or correct.