Create Exercise-7-RockPaperScissors
In the game Rock Paper Scissors, two players simultaneously choose one of three
options: rock, paper, or scissors. If both players choose the same option, then
the result is a tie. However, if they choose differently, the winner is
determined as follows:
  • Rock beats scissors, because a rock can break a pair of scissors.
  • Scissors beats paper, because scissors can cut paper.
  • Paper beats rock, because a piece of paper can cover a rock.
Create a game in which the computer randomly chooses rock, paper, or scissors.
Let the user enter a character, r, p, or s, each representing one of the three
choices. Then, determine the winner.